# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Movie Ticket Booking System API
  description: |
    A comprehensive REST API for managing movie ticket bookings, theatres, shows, and user authentication.

    This API provides:
    - User authentication and authorization
    - Movie and theatre management
    - Show scheduling and seat management
    - Transaction-safe booking system with stored procedures
    - Comprehensive error handling and validation

    **Key Features:**
    - JWT-based authentication
    - Role-based access control (USER, ADMIN)
    - Real-time seat availability
    - Secure payment processing
    - Database-first architecture with MySQL
  version: 1.0.0
  contact:
    name: Movie Booking System Support
    email: support@moviebooking.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.moviebooking.com/api
    description: Production server
paths:
...
  /bookings:
    get:
      tags:
        - Bookings
      summary: List all bookings (Admin only)
      description: Retrieve all bookings in the system with filtering options
      operationId: getAllBookings
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          description: Filter by user ID
          schema:
            type: integer
            minimum: 1
        - name: status
          in: query
          description: Filter by booking status
          schema:
            $ref: '#/components/schemas/BookingStatus'
        - name: dateFrom
          in: query
          description: Start date filter
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          description: End date filter
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/Page'
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 50
      responses:
        '200':
          description: Bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Bookings
      summary: Create booking (CRITICAL)
      description: |
        Create a new ticket booking. This is the most critical endpoint that:
        - Uses stored procedures for transaction safety
        - Implements seat locking to prevent double booking
        - Calculates amount server-side for security
        - Processes payment in the same transaction
      operationId: createBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingInput'
            example:
              showId: 1
              seatIds: [1, 2, 3]
              paymentMethod: "CREDIT_CARD"
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Invalid seat selection"
                  details: ["Seat A1 is not valid for this show"]
                  path: "/api/bookings"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Booking conflict (seats already booked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "BOOKING_CONFLICT"
                  message: "One or more seats are already booked"
                  details: ["Seat A1 is already booked", "Seat A2 is already booked"]
                  path: "/api/bookings"
        '422':
          description: Payment processing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bookings/my-bookings:
    get:
      tags:
        - Bookings
      summary: Get user's bookings
      description: Retrieve bookings for the authenticated user
      operationId: getUserBookings
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by booking status
          schema:
            $ref: '#/components/schemas/BookingStatus'
        - $ref: '#/components/parameters/Page'
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: User bookings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBookingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /bookings/{id}:
    get:
      tags:
        - Bookings
      summary: Get booking details
      description: |
        Retrieve detailed information about a specific booking.
        Users can only access their own bookings; admins can access all bookings.
      operationId: getBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Booking details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /bookings/{id}/cancel:
    put:
      tags:
        - Bookings
      summary: Cancel booking
      description: |
        Cancel an existing booking. Uses stored procedure for transaction safety.
        Business rules:
        - Can only cancel up to 2 hours before show time
        - Refund amount calculated based on cancellation policy
        - Seat availability automatically updated
      operationId: cancelBooking
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BookingId'
      responses:
        '200':
          description: Booking cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelBookingResponse'
        '400':
          description: Booking cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "CANCELLATION_NOT_ALLOWED"
                  message: "Booking cannot be cancelled within 2 hours of show time"
                  path: "/api/bookings/123/cancel"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Booking already cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  parameters:
...
    BookingId:
      name: id
      in: path
      required: true
      description: Booking ID
      schema:
        type: integer
        minimum: 1
...
    Page:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
  schemas:
...
    BookingStatus:
      type: string
      enum:
        - CONFIRMED
        - CANCELLED
    PaymentMethod:
      type: string
      enum:
        - CREDIT_CARD
        - DEBIT_CARD
        - UPI
        - NET_BANKING
    PaymentStatus:
      type: string
      enum:
        - PENDING
        - SUCCESS
        - FAILED
        - REFUNDED
    BookingInput:
      type: object
      required:
        - showId
        - seatIds
        - paymentMethod
      properties:
        showId:
          type: integer
          minimum: 1
          description: ID of the show to book
        seatIds:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: integer
            minimum: 1
          description: Array of seat IDs to book
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
    BookingSeat:
      type: object
      properties:
        seat_number:
          type: string
        seat_type:
          type: string
    Payment:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
          format: decimal
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        transaction_id:
          type: string
        payment_time:
          type: string
          format: date-time
    Booking:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        show_id:
          type: integer
        booking_time:
          type: string
          format: date-time
        total_amount:
          type: number
          format: decimal
        status:
          $ref: '#/components/schemas/BookingStatus'
        show:
          type: object
          properties:
            movie_title:
              type: string
            show_time:
              type: string
              format: date-time
            theatre_name:
              type: string
        seats:
          type: array
          items:
            $ref: '#/components/schemas/BookingSeat'
        payment:
          $ref: '#/components/schemas/Payment'
    BookingDetail:
      allOf:
        - $ref: '#/components/schemas/Booking'
        - type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: integer
                username:
                  type: string
                email:
                  type: string
    UserBooking:
      type: object
      properties:
        id:
          type: integer
        show_id:
          type: integer
        booking_time:
          type: string
          format: date-time
        total_amount:
          type: number
          format: decimal
        status:
          $ref: '#/components/schemas/BookingStatus'
        show:
          type: object
          properties:
            movie_title:
              type: string
            show_time:
              type: string
              format: date-time
            theatre_name:
              type: string
            screen_name:
              type: string
        seats_count:
          type: integer
        seats:
          type: array
          items:
            $ref: '#/components/schemas/BookingSeat'
    CancelledBooking:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [CANCELLED]
        cancelled_at:
          type: string
          format: date-time
        refund_amount:
          type: number
          format: decimal
        cancellation_fee:
          type: number
          format: decimal
    BookingResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            booking:
              $ref: '#/components/schemas/Booking'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'
    BookingDetailResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            booking:
              $ref: '#/components/schemas/BookingDetail'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'
    BookingsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            bookings:
              type: array
              items:
                $ref: '#/components/schemas/Booking'
            pagination:
              $ref: '#/components/schemas/Pagination'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'
    UserBookingsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            bookings:
              type: array
              items:
                $ref: '#/components/schemas/UserBooking'
            pagination:
              $ref: '#/components/schemas/Pagination'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'
    CancelBookingResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            booking:
              $ref: '#/components/schemas/CancelledBooking'
        message:
          type: string
        meta:
          $ref: '#/components/schemas/MetaData'

    # Utility Schemas
    MetaData:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        api_version:
          type: string
          example: "1.0"
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        hasNext:
          type: boolean
          description: Whether there's a next page
        hasPrev:
          type: boolean
          description: Whether there's a previous page
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human readable error message
            details:
              type: array
              items:
                type: string
              description: Additional error details
            path:
              type: string
              description: API endpoint path
        meta:
          $ref: '#/components/schemas/MetaData'
  responses:
...
    Unauthorized:
      description: Unauthorized - Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "AUTHENTICATION_ERROR"
              message: "Authentication required"
              path: "/api/bookings"
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "AUTHORIZATION_ERROR"
              message: "Admin access required"
              path: "/api/movies"
    NotFound:
      description: Not Found - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "RESOURCE_NOT_FOUND"
              message: "Movie not found"
              path: "/api/movies/999"
...
security:
  - bearerAuth: []
...